requirements.txt
fastapi
uvicorn
langchain
transformers
sentence-transformers
python-dotenv
gitpython
pydantic
PyGithub
qdrant-client
